Fix continuing from an app that returns more result
A real app (unlike the previous mocking code) cannot be drained
until the resolve() for the current queue item has been called,
because the ThingTalk code is blocked waiting for the queue item
to be processed.

Fix that, and improve the mocking code to more closely match
the behavior of thingengine-core.