dispatcher: implement a stack based dispatch mechanism
Implement explicit stack frames, which can either accept a message,
handle it internally (for control messages), or cause it to bubble
up the stack. The actual stack is the JS stack, which might or
might not be a good idea, but at least it requires minimal changes
to the dialog code.

Tests change minimally, mostly because the new code is stricter
with context handling.