Begin implementing the new stack-based dispatch mechanism
Dialog functions are no longer standalone functions, they are
part of the .dispatch() method in a stack frame. Additionally, dialogs
cannot interact with the user directly anymore. Instead, they must
a) push an inner stack frame, wait for it to complete, and then process
the result, or b) mark the command as compatible and handle it in
dispatch()